Refactor for Go 1.11 modules
