+++ go.mod
@@ -1,0 +1,11 @@
+module github.com/umeat/go-ntrip
+
+require (
+	github.com/aws/aws-sdk-go v1.17.1
+	github.com/benburkert/http v0.0.0-20140610191404-7b1c7d0a482c
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/geoscienceaustralia/go-rtcm v0.0.0-20190220051918-c57311901c51 // indirect
+	github.com/micro/go-config v0.14.0
+	github.com/satori/go.uuid v1.2.0
+	github.com/sirupsen/logrus v1.3.0
+)
