+++ ntrip/caster/types.go
@@ -1,4 +1,4 @@
-package ntrip
+package caster
 
 import (
     "errors"
